+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2026-05-23
+
 ### Added
 
 - **`LithairServerBuilder::with_handler` now wires the builder-level SSE